Removed the XNextEvent part in vo create window.
On multi-workspace window managers, when mplayer is spawned on a workspace
not visible currently, it doesn't play or respond to other commands, say in slave mode.
After running ltrace, mplayer was spinning at XNextEvent waiting for a MapNotify event.
Now window managers won't send this event until and unless the workspace is visible.
So changing the vo create function to not spin there. Tested with vdpau,gl VOs.